disasm: address tweaks

+++ src/disasm.h
@@ -3,6 +3,6 @@
 
 #include <stdint.h>
 
-void disasm(const uint8_t *buffer, const int size, const uint16_t address, const int illegal);
+void disasm(const uint8_t *buffer, const int size, const int address, const int illegal);
 
 #endif // DISASM_H_
